requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
cssselect==1.2.0
decorator==5.1.1
Django==5.0.6
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap5==24.2
django-crontab==0.7.1
django-jquery==3.1.0
django-stubs==5.0.2
django-stubs-ext==5.0.2
executing==2.0.1
idna==3.7
ipython==8.26.0
iquery==0.5.0
jedi==0.19.1
lxml==5.2.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prettytable==3.10.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyquery==2.0.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
setuptools==70.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
types-PyYAML==6.0.12.20240311
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
wheel==0.43.0
